Fix std::modff → std::modf (missed in previous CI fix)
Same issue as the other f-suffix math calls: std::modff is not a member
of std:: in strict C++ mode. Missed by the previous sed pass because
modf was not in the substitution list.

+++ lib/interpreter/glsl_std_450.cpp
@@ -589,7 +589,7 @@ Value dispatch_glsl_std_450(uint32_t inst_id, const std::vector<Value>& args,
         case Modf: {
             const Value& x = get(0);
             auto do_modf = [&](float v) -> std::pair<float, float> {
-                float i; float f = std::modff(v, &i); return {f, i};
+                float i; float f = std::modf(v, &i); return {f, i};
             };
             auto do_modf64 = [&](double v) -> std::pair<double, double> {
                 double i; double f = std::modf(v, &i); return {f, i};
